backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
pydantic==2.5.2
pydantic-settings==2.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
anthropic>=0.39.0
openai>=1.0.0
